// lib/components/Nav/Nav.Styles.d.ts
export declare const NavMenuItemWrapperStyles: {
    menuWrapper: string;
    subMenuWrapper: string;
};
export declare const navMenuItemStyles: {
    navItemWrapper: string;
    navItem: string;
    navParentItem: string;
    navArrow: string;
    navParentArrow: string;
    subMenuItem: string;
    innerSubNavToggler: string;
};